requirements.txt
llama-index
fastapi
python-dotenv
llama-index-llms-groq
llama-index-embeddings-huggingface
uvicorn
llama-cloud-services